refactor: extract shared useDerivedUsageData hook
Deduplicate the daily/sources/models derivation logic that was copied
between useUsageData and useUserProfile into a single reusable hook.
